distsqlrun: generalize mergejoiner tests
Factor out the merge joiner test cases into a general joiner test case.
Now the {hash,merge}joiners run eachothers tests as well as their own
from the general joiner test case repository.

Release note: None
